Conflicts:
	ide/src/org/alice/ide/declarationseditor/TypeMenu.java
	ide/src/org/alice/ide/declarationseditor/type/views/FieldsToolPaletteCoreView.java


Former-commit-id: 46c43e6459f79c6a9f29c28c8059005883bc7c7a

+++ ide/src/org/lgna/story/Scale.java
@@ -86,14 +86,14 @@ public final class Scale {
 	}
 
 	public Double getLeftToRight() {
-		return (double)this.internal.x;
+		return this.internal.x;
 	}
 
 	public Double getBottomToTop() {
-		return (double)this.internal.y;
+		return this.internal.y;
 	}
 
 	public Double getFrontToBack() {
-		return (double)this.internal.z;
+		return this.internal.z;
 	}
 }
